backend/src/requirements.txt
annotated-types==0.5.0
ansi2html==1.8.0
anyio==4.0.0
attrs==23.1.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.6.2
Bottleneck
Brotli==1.0.9
brotlipy==0.7.0
cachetools==5.3.0
certifi
cffi
charset-normalizer
chroma-hnswlib==0.7.2
chromadb==0.4.8
click==8.1.6
colorama
coloredlogs==15.0.1
conda==22.11.0
conda-content-trust
conda-package-handling
cryptography
Cython
dash==2.11.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
db-dtypes==1.1.1
docopt==0.6.2
exceptiongroup==1.1.3
fastapi==0.99.1
filelock==3.12.3
Flask==2.2.5
Flask-Compress==1.13
Flask-Cors==3.0.10
flatbuffers==23.5.26
fsspec==2023.6.0
ftfy==6.1.1
future==0.18.3
gensim
google-api-core==2.11.0
google-auth==2.16.0
google-cloud==0.34.0
google-cloud-bigquery==3.5.0
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpcio==1.51.1
grpcio-status==1.51.1
gunicorn==20.1.0
h11==0.14.0
httptools==0.6.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna
importlib-metadata==6.0.0
importlib-resources==6.0.1
ingestion==0.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib
Markdown==3.4.4
MarkupSafe==2.1.2
mkl-fft==1.3.1
mkl-random
mkl-service==2.4.0
monotonic==1.6
mpmath==1.3.0
nest-asyncio==1.5.7
networkx==3.1
nltk
numexpr
numpy==1.25.2
onnxruntime==1.15.1
open-clip-torch==2.20.0
outcome==1.2.0
overrides==7.4.0
packaging
pandas==1.4.4
Pillow==9.5.0
pipreqs==0.4.11
plotly==5.15.0
pluggy
posthog==3.0.2
proto-plus==1.22.2
protobuf==3.20.3
pulsar-client==3.3.0
pyarrow==12.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser
pydantic==1.10.12
pydantic_core==2.6.3
pyOpenSSL
pyparsing
PyPika==0.48.9
PySocks
python-dateutil
python-dotenv==1.0.0
python-multipart==0.0.6
pytz
PyYAML==6.0.1
regex
requests==2.31.0
retrying==1.3.4
rsa==4.9
ruamel-yaml-conda
ruamel.yaml
ruamel.yaml.clib
safetensors==0.3.3
scikit-learn
scipy==1.9.3
selenium==4.11.2
sentencepiece==0.1.99
six
smart-open
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
starlette==0.27.0
sympy==1.12
tenacity==8.2.2
threadpoolctl
timm==0.9.6
tokenizers==0.13.3
toolz
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.7.1
urllib3
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
wcwidth==0.2.6
websockets==11.0.3
Werkzeug==2.2.3
wsproto==1.2.0
xgboost==1.5.0
yarg==0.1.9
zipp==3.15.0
